+++ kie-server-parent/kie-server-services/kie-server-services-prometheus/src/main/java/org/kie/server/services/prometheus/PrometheusKieServerExtension.java
@@ -6,7 +6,6 @@ import java.util.Map;
 import java.util.ServiceLoader;
 
 import io.prometheus.client.CollectorRegistry;
-import io.prometheus.client.Histogram;
 import org.kie.server.api.KieServerConstants;
 import org.kie.server.api.model.Message;
 import org.kie.server.api.model.Severity;
@@ -125,41 +124,6 @@ public class PrometheusKieServerExtension implements KieServerExtension {
             messages.add(new Message(Severity.INFO, getExtensionName() + " is alive"));
         }
         return messages;
-    }
-
-    public void recordMetric() {
-
-        logger.info("Recording metrics: " + registry.hashCode());
-
-        Histogram histogram = Histogram.build().name("dmn_evaluate_decision_nanosecond" + System.nanoTime())
-                .help("DMN Evaluation Time")
-                .labelNames("decision_name")
-                .buckets(HALF_SECOND_NANO, toNano(1), toNano(2), toNano(3), toNano(4))
-                .register();
-
-        for(int i = 0; i < 10; i++){
-            int amt = 123456789 + i;
-            histogram.labels("prova")
-                    .observe(amt);
-
-            logger.info("inserted = " + amt);
-            try {
-                Thread.sleep(1000);
-            } catch (InterruptedException e) {
-                e.printStackTrace();
-            }
-        }
-
-    }
-
-    /**
-     * Number of nanoseconds in a second.
-     */
-    public static final long NANOSECONDS_PER_SECOND = 1_000_000_000;
-    public static final long HALF_SECOND_NANO = 500_000_000;
-
-    public static long toNano(long second) {
-        return second * NANOSECONDS_PER_SECOND;
     }
 
 }
