+++ kie-server-parent/kie-server-tests/kie-server-integ-tests-common/src/test/java/org/kie/server/integrationtests/common/rest/RestPrometheusIntegrationTest.java
@@ -15,10 +15,11 @@
 
 package org.kie.server.integrationtests.common.rest;
 
-import java.util.ServiceLoader;
 import javax.ws.rs.client.WebTarget;
 import javax.ws.rs.core.Response;
 
+import io.prometheus.client.CollectorRegistry;
+import io.prometheus.client.Histogram;
 import org.junit.Assert;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
@@ -28,7 +29,6 @@ import org.kie.server.api.model.ReleaseId;
 import org.kie.server.integrationtests.category.RESTOnly;
 import org.kie.server.integrationtests.config.TestConfig;
 import org.kie.server.integrationtests.shared.basetests.RestOnlyBaseIntegrationTest;
-import org.kie.server.services.prometheus.PrometheusKieServerExtension;
 
 @Category(RESTOnly.class)
 public class RestPrometheusIntegrationTest extends RestOnlyBaseIntegrationTest {
@@ -56,11 +56,7 @@ public class RestPrometheusIntegrationTest extends RestOnlyBaseIntegrationTest {
         System.setProperty(KieServerConstants.KIE_PROMETHEUS_SERVER_EXT_DISABLED, "false");
         KieContainerResource resource = new KieContainerResource("container", RELEASE_ID_1);
 
-        ServiceLoader<PrometheusKieServerExtension> appComponentsServices = ServiceLoader.load(PrometheusKieServerExtension.class);
-        for (PrometheusKieServerExtension extension : appComponentsServices) {
-            logger.info("extension = " + extension);
-            extension.recordMetric();
-        }
+        recordMetric();
 
         logger.info(configuration.getUserName());
         logger.info(configuration.getPassword());
@@ -71,10 +67,9 @@ public class RestPrometheusIntegrationTest extends RestOnlyBaseIntegrationTest {
 
             Assert.assertEquals(Response.Status.OK.getStatusCode(), response.getStatus());
 
-            Object res = response.readEntity(String.class);
-            Assert.assertEquals("pippo", res);
-
-
+            String res = response.readEntity(String.class);
+            String[] split = res.split("\\n");
+            Assert.assertEquals(10, split.length);
         } catch (Exception e) {
             throw new RuntimeException(
                     "Unexpected exception creating container: " + resource.getContainerId() + " with release-id " + resource.getReleaseId(),
@@ -84,5 +79,34 @@ public class RestPrometheusIntegrationTest extends RestOnlyBaseIntegrationTest {
                 response.close();
             }
         }
+    }
+
+    public void recordMetric() {
+
+        CollectorRegistry registry = CollectorRegistry.defaultRegistry;
+
+        logger.info("Recording metrics: " + registry.hashCode());
+
+        Histogram histogram = Histogram.build().name("dmn_evaluate_decision_nanosecond" + System.nanoTime())
+                .help("DMN Evaluation Time")
+                .labelNames("decision_name")
+                .buckets(HALF_SECOND_NANO, toNano(1), toNano(2), toNano(3), toNano(4))
+                .register();
+
+        int amt = 123456789 ;
+        histogram.labels("prova")
+                .observe(amt);
+
+        logger.info("inserted = " + amt);
+    }
+
+    /**
+     * Number of nanoseconds in a second.
+     */
+    public static final long NANOSECONDS_PER_SECOND = 1_000_000_000;
+    public static final long HALF_SECOND_NANO = 500_000_000;
+
+    public static long toNano(long second) {
+        return second * NANOSECONDS_PER_SECOND;
     }
 }
